Implement bitcode attaching and extraction
This change attaches bitcode to object files as they are created.  This commit
also adds a subcommand to extract bitcode from results.

Bitcode files are wrapped in a tar archive when they are attached to object
files (in their own special .llvm_bitcode section).  Since generic data sections
like that are concatenated at link time, the tar files are concatenated.  It
turns out that concatenated tar files are still valid tar files, so we can
extract all of the bitcode after linking.