Add integration tests (#11)
The test builds bitcode for zlib and ensures that the bitcode can be extracted
from the shared library. It does not ensure that the bitcode is correct yet.

The integration test implementation needs to be refactored to promote code
reuse, but I'd rather get it working before doing that.

This ended up requiring a few related changes. These could have been extracted
into smaller individual commits, but it probably isn't worth the effort.  The
changes are:

- Improved the module structure to expose the logic of `main` as a function that
can be reused from the tests
- Fixed the module structure (the original top-level module was incorrect and
also never built); this includes necessary updates to Cargo.toml
- Compute an exit code from the top-level build command and propagates it to the
executable (useful in general, but especially helpful for the test suite to
detect if the build fails unexpectedly)
- Modify the extract command to take the path to `llvm-link` directly, rather
than an executable suffix (this is more flexible, mirrors other tools, and makes
the test harness simpler)