Echo last file captured into (helpful for post-capture extraction).
Frequently the build operation performed during the `generate-bitcode`
sequence will build various dependencies and then a final library or
executable.  This final file usually contains the summed LLVM IR
bitcode capture for the entire process and is the most likely target
of a subsequent `extract-bitcode` operation.  This is not always the
case (e.g. a "test" build will probably build a testing executable as
the last step), but that can usually be addressed by adjusting the
build command passed to the `generate-bitcode` operation.

By echoing this last file to stdout along with the other summary
stats, various automation tooling wishing to extract and consume the
bitcode will have a convenient identification of the most likely file
of interest.