Always supply -g and unless --flags-unchanged, ensure -O0 for bitcode-generation
When generating bitcode, the -g flag is always supplied to add debug
information to the output bitcode file.  The -O0 flag will be
specified as well to suppress optimization, unless the new
--flags-unchanged/-f argument is specified, in which case the
optimization setting (if any) of the original compilation command will
be passed to the bitcode generation as well.