Automatically generate debug info when building bitcode
This is almost always useful, but also provide a new command line flag,
`--suppress-automatic-debug`, to avoid it if needed (e.g., if the generated
debug info is too large to be practical).

This commit includes a minor refactoring to bundle together all of the options
that control bitcode generation into a single structure that can be passed more
easily throughout the bitcode generation process.