Support skipping blacklisted arguments with values.
If a blacklisted argument takes a value, this adds support for
skipping the value as well as the argument.

For user-supplied rejected arguments, this gives a best-shot if the
argument happens to be recognized, but if the user skips a unary
argument that is unknown to build-bom, that argument will probably
appear as an extra input file.