Add end-of-regex anchors for matching compile commands.
This one was fun to diagnose.  The autoconf/automake build environment
was specifying the build as:

  .c.lo:
    $(LTCOMPILE) -MT $@ -MD -MP -MF $$depbase.Tpo -c -o $@ $<

where

  LTCOMPILE = $(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) \
    $(LIBTOOLFLAGS) --mode=compile $(CC) $(DEFS) \
    $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) \
    $(AM_CFLAGS) $(CFLAGS)

and build-bom was failing because it was detecting multiple input
files from a command line looking like:

  gcc -c -g -I. -I../.. -I ../../src -MD .libs/compat_strlcat.d -MF .deps/compat_strlcat.Tpo -MP -MT compat_strlcat.lo -D _FORTIFY_SOURCE=2 -D HAVE_CONFIG_H [..lots more include args and such...] compat_strlcat.c -mtune=generic -march=x86-64 -g -O2 -O2 -Wformat=1 -Wformat-security -Wall -Wextra -Wno-unused-parameter -Wstrict-aliasing=2 -Wno-error -fPIC -fstack-protector-strong -fno-strict-overflow -frandom-seed=73nic52akm -fcommon -o compat_strlcat.lo

The two input file arguments identified by build-bom were:
  .libs/compat_strlcat.d and compat_strlcat.c

Checking the GCC documentation, the -MD argument has no value (it is a
nullary argument, not a unary argument).

After lots of debugging, crawling through autoconf/automake output,
libtool code, etc., I finally had a suspicion and verified it by
adding a `-v` flag to the LTCOMPILE for the `.c.lo:` rule, and sure
enough:

  ...
  gcc version 10.3.0 (GCC)
  collect_gcc_options='-Wformat=1' ... '-MT' 'compat_strlcat.lo' '-MD' '-MP' '-MF' ...
  ...
  cc1 -quiet -v -I . -I ../.. -I ../../src -MD .libs/compat_getpass.d -MF .deps/compat_getpass.Tpo -MP -MT compat_getpass.lo ...

Where build-bom was catching the cc1 invocation as a compile command
because it was matching the command against r"cc" (among others) which
matched "cc1" just fine.

The fix here is to anchor the regex-matching of compile commands to
the end of the command name.