Enhance output file detection for -oFILE as well as -o FILE.
The `-o` option is a single-letter option and allows the argument to immediately
follow the option.  The target file needs to be identified by build-bom in order
to determine where to place the generated bitcode, so it needs to handle both of
these cases.

Also added some unit tests to verify this behavior.