Add some additional blacklisted arguments.
Of particular interest, the original compilation may be a cross-compiler
targeting a non-native platform.  While LLVM is not truly target-independent, it
is mostly so, and probably sufficiently so for current intended users.  Passing
through the target specification arguments might cause a failure unless the llvm
installation had equivalent cross-compilation targets, so this change includes
removing those arguments via blacklisting.