Add chainsop module for chained operations.
The build-bom functionality largely consists of issuing a sequence of commands as
sub-processes at key points in the processing.  When running the sequence of
operations, the output of one operation is usually the input of the subsequent
operation; sometimes temporary files are needed for these interspersed outputs.

This extracts the functionality of managing subprocess execution, temporary file
creation, and input-to-output chaining into the chainops module.  This simplifies
the mainline code and should facilitate future modifications of the chained
sequence of operations.