Remove -b/--bc-out option: no longer needed.
With the change to use the chainsop that automatically provides temporary files,
the bitcode files are no longer being built "next" to the object files, and
therefore the scenario that generated the addition of this functionality is no
longer applicable.

This flag is actually ignored in the chainsop-adding patch, but the official
removal of it is done here in a separate patch to allow it to be easily
re-introduced should a viable need for it arise in the future.