Fix bitcode operations for lack of an explicit output file. (#84)
* Fix bitcode operations for lack of an explicit output file.

If the compile command did not specify an explicit output file (`cc ... -o
outfile ...`) then build-bom infers one from the input filename.  However, if
there is a chain of multiple operations, that inferred name should be added to
the operation that generates the bitcode output file.  This was previously
missing, causing the attempt to generate the bitcode file for inclusion into the
resulting ELF file to fail.

This also changes the severity (and wording) of the output message to more
appropriately indicate when there are actual problems generating the bitcode and
not just suspicious situations.

* Fix formatting.